Move homepage items filtering to a separate file
